[ refactor ] SocialLoginBtn component를 LoginPage 폴더로 이동
